// src/app/module-manager/models/application.ts
export class Application {
    uuid: string ="";
    code: string ="";
    nom: string ="";
    version: string ="";
    environnement: string ="";
    description: string ="";
    url: string ="";
    dependances: string ="";
    enabled: boolean = false;
    icon: string = "";
    color: string = "";
}